+++ lib/unisec.rb
@@ -3,10 +3,12 @@
 require 'unisec/version'
 
 require 'unisec/bidi'
+require 'unisec/blocks'
 require 'unisec/confusables'
 require 'unisec/decdump'
 require 'unisec/hexdump'
 require 'unisec/normalization'
+require 'unisec/planes'
 require 'unisec/properties'
 require 'unisec/rugrep'
 require 'unisec/size'